Functional dynamic settings, with settings saving and command-line support
Saved settings from previous runs only replace the default settings, and command-line settings skip requesting input for that setting entirely (effectively overriding everything).
This should be further refactored to use a map between or in place of the settings structure and everything else, which should remove most of the duplicate places to maintain te list of settings and their types.
This should additionally be refactored to replace the settings_requesting_guide json first-class data type with a raw string literal with deserialization; the readability of the first-class is abysmal.